Move endpoints to audius-backend/endpoints
