+++ packages/web/src/services/audius-backend/endpoints.ts
@@ -1,0 +1,3 @@
+export const IDENTITY_SERVICE = process.env.REACT_APP_IDENTITY_SERVICE
+export const USER_NODE = process.env.REACT_APP_USER_NODE
+export const LEGACY_USER_NODE = process.env.REACT_APP_LEGACY_USER_NODE
